# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk layanan MongoDB pada aplikasi Go Fiber.",
        "title": "Go Fiber Mongo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/go-fiber-mongo",
    "paths": {
        "/alumni": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar alumni dengan pagination, sorting, dan pencarian",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Daftar alumni",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah per halaman",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kolom sortir",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Urutan asc/desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kata kunci pencarian",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAllAlumniResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat data alumni baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Tambah alumni",
                "parameters": [
                    {
                        "description": "Permintaan pembuatan alumni",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.CreateAlumniRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreateAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreateAlumniResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreateAlumniResponse"
                        }
                    }
                }
            }
        },
        "/alumni/check": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengecek status alumni menggunakan API key legacy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Cek alumni berdasarkan NIM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NIM Mahasiswa",
                        "name": "nim",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.CheckAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/alumni/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data alumni berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Detail alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAlumniByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAlumniByIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAlumniByIDResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data alumni berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Perbarui alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permintaan pembaruan alumni",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data alumni berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Hapus alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteAlumniResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteAlumniResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Mengembalikan token JWT untuk mengakses API Mongo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth (Mongo)"
                ],
                "summary": "Login user (Mongo)",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
        "/pekerjaan/alumni/{alumni_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil seluruh pekerjaan milik alumni tertentu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerjaan (Mongo)"
                ],
                "summary": "Daftar pekerjaan per alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "alumni_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByAlumniIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByAlumniIDResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByAlumniIDResponse"
                        }
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil profil user berdasarkan token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile (Mongo)"
                ],
                "summary": "Profil pengguna saat ini",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil seluruh role yang tersedia",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles (Mongo)"
                ],
                "summary": "Daftar role",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.ListRolesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat role baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles (Mongo)"
                ],
                "summary": "Tambah role",
                "parameters": [
                    {
                        "description": "Permintaan pembuatan role",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.CreateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreateRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreateRoleResponse"
                        }
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data role berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles (Mongo)"
                ],
                "summary": "Detail role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Role",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetRoleByIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetRoleByIDResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data role berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles (Mongo)"
                ],
                "summary": "Perbarui role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Role",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permintaan pembaruan role",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateRoleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateRoleResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus role berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles (Mongo)"
                ],
                "summary": "Hapus role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Role",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteRoleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteRoleResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
...
        "mongo.CheckAlumniResponse": {
            "type": "object",
            "properties": {
                "alumni": {
                    "$ref": "#/definitions/mongo.Alumni"
                },
                "isAlumni": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "mongo.CreateAlumniRequest": {
            "type": "object",
            "required": [
                "angkatan",
                "email",
                "jurusan",
                "nama",
                "nim",
                "password",
                "role_id",
                "tahun_lulus"
            ],
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "angkatan": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "no_telepon": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "integer"
                }
            }
        },
        "mongo.CreateAlumniResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.Alumni"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "mongo.DeleteAlumniResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "mongo.GetAllAlumniResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.AlumniData"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "mongo.GetAlumniByIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.Alumni"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "mongo.UpdateAlumniRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "angkatan": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "no_telepon": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}